A cleaner approach to pause/unpause
The previous approach used `playdate.stop` to cease calls to update,
which also required jumping through some hoops to force the debug
layer to redraw on any updates, and felt a bit like a hack.

The new approach avoid that, and instead usurps the `playdate.update`
function directly, which means that the debug layer still updates in
real time even though the underlying game is paused. This avoids the
hack and will enable animated effects in the debug layer.