Add support for custom draw offsets
Acetate debug drawing needs to align properly with sprites. To do so,
Acetate adjusts the draw offset for the drawing context before calling
debugDraw on each sprite, but it lacks information about any adjustments
to this offset made by the application. This change both provides
better out-of-the-box behavior for apps which use custom draw offsets,
and a means to manually cache the offset for a given sprite as needed
to enable proper debug drawing.